+++ src/app/page.tsx
@@ -18,7 +18,8 @@ export default async function Home() {
           <SignInButton />
         </Button>
       </SignedOut>
-      {sessionClaims?.role}
+      {sessionClaims?.roles?.includes("admin") && <h1>Admin</h1>}
+      {sessionClaims?.roles}
       <div className="container flex max-w-[64rem] flex-col items-center gap-4 text-center">
         <Icons.logo className="h-16 w-16" />
         <h1 className="text-4xl font-semibold sm:text-5xl md:text-6xl lg:text-7xl">
