+++ src/types/types.ts
@@ -1,13 +1,20 @@
 // Templated Role Based Access Control
 // Source: https://www.youtube.com/watch?v=5GG-VUvruzE&t=180s
-type Comment = {
+
+import {
+  PermissionCheck,
+  Permissions,
+  Role,
+} from "@/lib/auth/role-based-access-control"
+
+export type Comment = {
   id: string
   body: string
   authorId: string
   createdAt: Date
 }
 
-type Todo = {
+export type Todo = {
   id: string
   title: string
   userId: string
@@ -15,99 +22,10 @@ type Todo = {
   invitedUsers: string[]
 }
 
-type Role = "admin" | "moderator" | "user"
-type User = { blockedBy: string[]; roles: Role[]; id: string }
-
-type PermissionCheck<Key extends keyof Permissions> =
-  | boolean
-  | ((user: User, data: Permissions[Key]["dataType"]) => boolean)
-
-type RolesWithPermissions = {
+export type RolesWithPermissions = {
   [R in Role]: Partial<{
     [Key in keyof Permissions]: Partial<{
       [Action in Permissions[Key]["action"]]: PermissionCheck<Key>
     }>
   }>
 }
-
-type Permissions = {
-  comments: {
-    dataType: Comment
-    action: "view" | "create" | "update"
-  }
-  todos: {
-    // Can do something like Pick<Todo, "userId"> to get just the rows you use
-    dataType: Todo
-    action: "view" | "create" | "update" | "delete"
-  }
-}
-
-const ROLES = {
-  admin: {
-    comments: {
-      view: true,
-      create: true,
-      update: true,
-    },
-    todos: {
-      view: true,
-      create: true,
-      update: true,
-      delete: true,
-    },
-  },
-  moderator: {
-    comments: {
-      view: true,
-      create: true,
-      update: true,
-    },
-    todos: {
-      view: true,
-      create: true,
-      update: true,
-      delete: (user, todo) => todo.completed,
-    },
-  },
-  user: {
-    comments: {
-      view: (user, comment) => !user.blockedBy.includes(comment.authorId),
-      create: true,
-      update: (user, comment) => comment.authorId === user.id,
-    },
-    todos: {
-      view: (user, todo) => !user.blockedBy.includes(todo.userId),
-      create: true,
-      update: (user, todo) =>
-        todo.userId === user.id || todo.invitedUsers.includes(user.id),
-      delete: (user, todo) =>
-        (todo.userId === user.id || todo.invitedUsers.includes(user.id)) &&
-        todo.completed,
-    },
-  },
-} as const satisfies RolesWithPermissions
-
-export function hasPermission<Resource extends keyof Permissions>(
-  user: User,
-  resource: Resource,
-  action: Permissions[Resource]["action"],
-  data?: Permissions[Resource]["dataType"]
-) {
-  return user.roles.some((role) => {
-    const permission = (ROLES as RolesWithPermissions)[role][resource]?.[action]
-    if (permission == null) return false
-
-    if (typeof permission === "boolean") return permission
-    return data != null && permission(user, data)
-  })
-}
-
-// USAGE:
-// Can create a comment
-// hasPermission(user, "comments", "create")
-
-// // Can view the `todo` Todo
-// hasPermission(user, "todos", "view", todo)
-
-// // Can view all todos
-// hasPermission(user, "todos", "view")
